Move functions related to `value_list_t` to a separate package.
