+++ src/daemon/plugin.c
@@ -673,56 +673,6 @@ static void stop_read_threads(void) {
   read_threads_num = 0;
 } /* void stop_read_threads */
 
-static void plugin_value_list_free(value_list_t *vl) /* {{{ */
-{
-  if (vl == NULL)
-    return;
-
-  meta_data_destroy(vl->meta);
-  sfree(vl->values);
-  sfree(vl);
-} /* }}} void plugin_value_list_free */
-
-static value_list_t *
-plugin_value_list_clone(value_list_t const *vl_orig) /* {{{ */
-{
-  value_list_t *vl;
-
-  if (vl_orig == NULL)
-    return NULL;
-
-  vl = malloc(sizeof(*vl));
-  if (vl == NULL)
-    return NULL;
-  memcpy(vl, vl_orig, sizeof(*vl));
-
-  if (vl->host[0] == 0)
-    sstrncpy(vl->host, hostname_g, sizeof(vl->host));
-
-  vl->values = calloc(vl_orig->values_len, sizeof(*vl->values));
-  if (vl->values == NULL) {
-    plugin_value_list_free(vl);
-    return NULL;
-  }
-  memcpy(vl->values, vl_orig->values,
-         vl_orig->values_len * sizeof(*vl->values));
-
-  vl->meta = meta_data_clone(vl->meta);
-  if ((vl_orig->meta != NULL) && (vl->meta == NULL)) {
-    plugin_value_list_free(vl);
-    return NULL;
-  }
-
-  if (vl->time == 0)
-    vl->time = cdtime();
-
-  /* Fill in the interval from the thread context, if it is zero. */
-  if (vl->interval == 0)
-    vl->interval = plugin_get_interval();
-
-  return vl;
-} /* }}} value_list_t *plugin_value_list_clone */
-
 static void write_queue_ref_single(write_queue_elem_t *elem, long dir) {
   elem->ref_count += dir;
 
@@ -2249,104 +2199,6 @@ EXPORT int plugin_dispatch_metric_family(metric_family_t const *fam) {
 
   return status;
 }
-
-EXPORT int plugin_dispatch_values(value_list_t const *vl) {
-  data_set_t const *ds = plugin_get_ds(vl->type);
-  if (ds == NULL) {
-    return EINVAL;
-  }
-
-  for (size_t i = 0; i < vl->values_len; i++) {
-    metric_family_t *fam = plugin_value_list_to_metric_family(vl, ds, i);
-    if (fam == NULL) {
-      int status = errno;
-      ERROR("plugin_dispatch_values: plugin_value_list_to_metric_family "
-            "failed: %s",
-            STRERROR(status));
-      return status;
-    }
-
-    int status = plugin_dispatch_metric_family(fam);
-    metric_family_free(fam);
-    if (status != 0) {
-      return status;
-    }
-  }
-
-  return 0;
-}
-
-__attribute__((sentinel)) int
-plugin_dispatch_multivalue(value_list_t const *template, /* {{{ */
-                           bool store_percentage, int store_type, ...) {
-  value_list_t *vl;
-  int failed = 0;
-  gauge_t sum = 0.0;
-  va_list ap;
-
-  assert(template->values_len == 1);
-
-  /* Calculate sum for Gauge to calculate percent if needed */
-  if (DS_TYPE_GAUGE == store_type) {
-    va_start(ap, store_type);
-    while (42) {
-      char const *name;
-      gauge_t value;
-
-      name = va_arg(ap, char const *);
-      if (name == NULL)
-        break;
-
-      value = va_arg(ap, gauge_t);
-      if (!isnan(value))
-        sum += value;
-    }
-    va_end(ap);
-  }
-
-  vl = plugin_value_list_clone(template);
-  /* plugin_value_list_clone makes sure vl->time is set to non-zero. */
-  if (store_percentage)
-    sstrncpy(vl->type, "percent", sizeof(vl->type));
-
-  va_start(ap, store_type);
-  while (42) {
-    char const *name;
-    int status;
-
-    /* Set the type instance. */
-    name = va_arg(ap, char const *);
-    if (name == NULL)
-      break;
-    sstrncpy(vl->type_instance, name, sizeof(vl->type_instance));
-
-    /* Set the value. */
-    switch (store_type) {
-    case DS_TYPE_GAUGE:
-      vl->values[0].gauge = va_arg(ap, gauge_t);
-      if (store_percentage)
-        vl->values[0].gauge *= sum ? (100.0 / sum) : NAN;
-      break;
-    case DS_TYPE_COUNTER:
-      vl->values[0].counter = va_arg(ap, counter_t);
-      break;
-    case DS_TYPE_DERIVE:
-      vl->values[0].derive = va_arg(ap, derive_t);
-      break;
-    default:
-      ERROR("plugin_dispatch_multivalue: given store_type is incorrect.");
-      failed++;
-    }
-
-    status = plugin_dispatch_values(vl);
-    if (status != 0)
-      failed++;
-  }
-  va_end(ap);
-
-  plugin_value_list_free(vl);
-  return failed;
-} /* }}} int plugin_dispatch_multivalue */
 
 EXPORT int plugin_dispatch_notification(const notification_t *notif) {
   llentry_t *le;
